FileCSSInjector
Added functionality for targeted injection of CSS values from a
source XD file's CSS to a site's theme-default.css that is
manually downloaded and placed in the work folder. This
should be improved upon in the future to so that the file is
moved automatically via SSH. This work is apart of #10.